Refactor: Service 하나로 통합, 컨트롤러 바로 결과 나오게 수정
